# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LearnMate AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /flashcard-decks:
    get:
      operationId: listFlashcardDecks
      tags: [flashcards]
      summary: List all flashcard decks
      responses:
        "200":
          description: List of flashcard decks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FlashcardDeck"
    post:
      operationId: createFlashcardDeck
      tags: [flashcards]
      summary: Create a new flashcard deck
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlashcardDeckInput"
      responses:
        "201":
          description: Created flashcard deck
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlashcardDeck"
  /flashcard-decks/{id}:
    get:
      operationId: getFlashcardDeck
      tags: [flashcards]
      summary: Get a flashcard deck by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Flashcard deck with cards
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlashcardDeckWithCards"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateFlashcardDeck
      tags: [flashcards]
      summary: Update a flashcard deck
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlashcardDeckUpdate"
      responses:
        "200":
          description: Updated flashcard deck
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlashcardDeck"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteFlashcardDeck
      tags: [flashcards]
      summary: Delete a flashcard deck
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /flashcard-decks/{id}/cards:
    get:
      operationId: listFlashcardDeckCards
      tags: [flashcards]
      summary: List all cards in a deck
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of flashcards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Flashcard"
    post:
      operationId: createFlashcardDeckCard
      tags: [flashcards]
      summary: Add a card to a deck
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlashcardInput"
      responses:
        "201":
          description: Created flashcard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Flashcard"
  /flashcard-decks/{id}/cards/{cardId}:
    patch:
      operationId: updateFlashcardDeckCard
      tags: [flashcards]
      summary: Update a flashcard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FlashcardUpdate"
      responses:
        "200":
          description: Updated flashcard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Flashcard"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteFlashcardDeckCard
      tags: [flashcards]
      summary: Delete a flashcard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: cardId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    FlashcardDeck:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        title:
          type: string
        subject:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        cardCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - subject
        - description
        - cardCount
        - createdAt
        - updatedAt
    FlashcardDeckWithCards:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        title:
          type: string
        subject:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        cardCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        cards:
          type: array
          items:
            $ref: "#/components/schemas/Flashcard"
      required:
        - id
        - userId
        - title
        - subject
        - description
        - cardCount
        - createdAt
        - updatedAt
        - cards
    FlashcardDeckInput:
      type: object
      properties:
        title:
          type: string
        subject:
          type: string
        description:
          type: string
      required:
        - title
    FlashcardDeckUpdate:
      type: object
      properties:
        title:
          type: string
        subject:
          type: string
        description:
          type: string
    Flashcard:
      type: object
      properties:
        id:
          type: integer
        deckId:
          type: integer
        front:
          type: string
        back:
          type: string
        difficulty:
          type: ["string", "null"]
        isBookmarked:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - deckId
        - front
        - back
        - difficulty
        - isBookmarked
        - createdAt
    FlashcardInput:
      type: object
      properties:
        front:
          type: string
        back:
          type: string
        difficulty:
          type: string
      required:
        - front
        - back
    FlashcardUpdate:
      type: object
      properties:
        front:
          type: string
        back:
          type: string
        difficulty:
          type: string
        isBookmarked:
          type: boolean
...